backend/requirements.txt
langgraph
langchain
pydantic
pytesseract
pdfplumber
pillow